01_etl/requirements.txt
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
elastic-transport==8.1.2
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
idna==3.3
ordereddict==1.1
packaging==21.3
psycopg2-binary==2.9
pydantic==1.9.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.2.0
requests==2.27.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1